SFA: Prepare to implement framebuffer effects
