+++ src/StarFoxAdventures/maps.ts
@@ -313,7 +313,23 @@ class MapSceneRenderer extends SFARenderer {
     }
     
     protected renderWorld(device: GfxDevice, renderInstManager: GfxRenderInstManager, viewerInput: Viewer.ViewerRenderInput) {
+        // Prolog
+        const template = this.renderHelper.pushTemplateRenderInst();
+        fillSceneParamsDataOnTemplate(template, viewerInput, false);
+
+        // Body
         this.map.prepareToRender(device, renderInstManager, viewerInput);
+
+        // Epilog
+        renderInstManager.popTemplateRenderInst();
+
+        let hostAccessPass = device.createHostAccessPass();
+        this.prepareToRender(device, hostAccessPass, viewerInput);
+        device.submitPass(hostAccessPass);
+        
+        renderInstManager.drawOnPassRenderer(device, this.renderPass);
+        renderInstManager.resetRenderInsts();
+        device.submitPass(this.renderPass);
     }
 }
