+++ src/StarFoxAdventures/render.ts
@@ -8,7 +8,7 @@ import { Camera, computeViewMatrix } from '../Camera';
 import ArrayBufferSlice from '../ArrayBufferSlice';
 import { GXMaterial } from '../gx/gx_material';
 import { GfxRenderCache } from '../gfx/render/GfxRenderCache';
-import { standardFullClearRenderPassDescriptor, noClearRenderPassDescriptor, BasicRenderTarget } from '../gfx/helpers/RenderTargetHelpers';
+import { standardFullClearRenderPassDescriptor, noClearRenderPassDescriptor, BasicRenderTarget, ColorTexture } from '../gfx/helpers/RenderTargetHelpers';
 
 import { SFATexture } from './textures';
 
@@ -36,6 +36,9 @@ export abstract class SFARendererHelper implements Viewer.SceneGfx {
 }
 
 export class SFARenderer extends SFARendererHelper {
+    protected renderPass: GfxRenderPass;
+    protected sceneTexture = new ColorTexture();
+
     protected renderSky(device: GfxDevice, renderInstManager: GfxRenderInstManager, viewerInput: Viewer.ViewerRenderInput) {}
 
     protected renderWorld(device: GfxDevice, renderInstManager: GfxRenderInstManager, viewerInput: Viewer.ViewerRenderInput) {}
@@ -47,41 +50,13 @@ export class SFARenderer extends SFARendererHelper {
     public render(device: GfxDevice, viewerInput: Viewer.ViewerRenderInput): GfxRenderPass {
         const renderInstManager = this.renderHelper.renderInstManager;
         this.renderTarget.setParameters(device, viewerInput.backbufferWidth, viewerInput.backbufferHeight);
+        this.sceneTexture.setParameters(device, viewerInput.backbufferWidth, viewerInput.backbufferHeight);
+        this.renderPass = this.renderTarget.createRenderPass(device, viewerInput.viewport, standardFullClearRenderPassDescriptor);
 
-        // Draw sky
-        const skyTemplate = this.renderHelper.pushTemplateRenderInst();
-        const oldProjection = mat4.create();
-        mat4.copy(oldProjection, viewerInput.camera.projectionMatrix);
-        mat4.identity(viewerInput.camera.projectionMatrix);
-        fillSceneParamsDataOnTemplate(skyTemplate, viewerInput, false);
         this.renderSky(device, renderInstManager, viewerInput);
-        renderInstManager.popTemplateRenderInst();
-
-        mat4.copy(viewerInput.camera.projectionMatrix, oldProjection);
-
-        let hostAccessPass = device.createHostAccessPass();
-        this.prepareToRender(device, hostAccessPass, viewerInput);
-        device.submitPass(hostAccessPass);
-        
-        let passRenderer = this.renderTarget.createRenderPass(device, viewerInput.viewport, standardFullClearRenderPassDescriptor);
-        renderInstManager.drawOnPassRenderer(device, passRenderer);
-        renderInstManager.resetRenderInsts();
-        device.submitPass(passRenderer);
-
-        // Draw world
-        const template = this.renderHelper.pushTemplateRenderInst();
-        fillSceneParamsDataOnTemplate(template, viewerInput, false);
         this.renderWorld(device, renderInstManager, viewerInput);
-        renderInstManager.popTemplateRenderInst();
 
-        hostAccessPass = device.createHostAccessPass();
-        this.prepareToRender(device, hostAccessPass, viewerInput);
-        device.submitPass(hostAccessPass);
-
-        passRenderer = this.renderTarget.createRenderPass(device, viewerInput.viewport, noClearRenderPassDescriptor);
-        renderInstManager.drawOnPassRenderer(device, passRenderer);
-        renderInstManager.resetRenderInsts();
-        // Pass will be submitted by the caller
-        return passRenderer;
+        this.renderPass = this.renderTarget.createRenderPass(device, viewerInput.viewport, noClearRenderPassDescriptor);
+        return this.renderPass;
     }
 }
